use inplace mode for WOQ
